+++ src/main/java/it/lpleo/adventofcode/y2019/p5/domain/HandlerOutput.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class HandlerOutput {
   private boolean wait;
-  private int result;
+  private long result;
 }
