+++ src/test/java/it/lpleo/adventofcode/y2019/p5/ChanceOfAsteroidsPuzzleSolverTest.java
@@ -42,82 +42,82 @@ public class ChanceOfAsteroidsPuzzleSolverTest {
 
   @Test
   public void solvePart1InputOutputTest() {
-    VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(new int[]{3, 0, 4, 0, 99});
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(6));
+    VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(new long[]{3, 0, 4, 0, 99});
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(6L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine,
         asList(new InputHandler(fifoQueue), new OutputHandler(fifoQueue), new ErrorHandler()));
 
-    assertThat(fifoQueue.poll(), is(6));
+    assertThat(fifoQueue.poll(), is(6L));
   }
 
   @Test
   public void solvePart1ParameterImmediateModeTest() {
-    VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(new int[]{1002, 4, 3, 4, 33});
+    VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(new long[]{1002, 4, 3, 4, 33});
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine,
         asList(new MultiplicationHandler(), new ErrorHandler()));
 
-    assertThat(vonNeumannMachine.getValue(4), is(99));
+    assertThat(vonNeumannMachine.getValue(4), is(99L));
   }
 
   @Test
   public void solvePart2IfEqualsTo8PositionModeOutput1() {
     VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(
-        new int[]{3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8});
+        new long[]{3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8});
 
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(8));
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(8L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine, getPT2Handlers(fifoQueue));
 
-    assertThat(fifoQueue.poll(), is(1));
+    assertThat(fifoQueue.poll(), is(1L));
   }
 
   @Test
   public void solvePart2IfLessThan8PositionModeOutput1() {
     VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(
-        new int[]{3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8});
+        new long[]{3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8});
 
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(5));
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(5L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine, getPT2Handlers(fifoQueue));
 
-    assertThat(fifoQueue.poll(), is(1));
+    assertThat(fifoQueue.poll(), is(1L));
   }
 
   @Test
   public void solvePart2IfEqualTo8ImmediateModeOutput1() {
     VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(
-        new int[]{3, 3, 1108, -1, 8, 3, 4, 3, 99});
+        new long[]{3, 3, 1108, -1, 8, 3, 4, 3, 99});
 
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(8));
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(8L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine, getPT2Handlers(fifoQueue));
 
-    assertThat(fifoQueue.poll(), is(1));
+    assertThat(fifoQueue.poll(), is(1L));
   }
 
   @Test
   public void solvePart2IfEqualsTo8ImmediateModeOutput1() {
     VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(
-        new int[]{3, 3, 1107, -1, 8, 3, 4, 3, 99});
+        new long[]{3, 3, 1107, -1, 8, 3, 4, 3, 99});
 
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(9));
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(9L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine, getPT2Handlers(fifoQueue));
 
-    assertThat(fifoQueue.poll(), is(0));
+    assertThat(fifoQueue.poll(), is(0L));
   }
 
   @Test
   public void solvePart2ComplexExample() {
     VonNeumannMachine vonNeumannMachine = new VonNeumannMachine(
-        new int[]{3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31,
+        new long[]{3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31,
             1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104,
             999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99});
 
-    LinkedList<Integer> fifoQueue = new LinkedList<>(singletonList(7));
+    LinkedList<Long> fifoQueue = new LinkedList<>(singletonList(7L));
     StoppableVonNeumannMachineRunner.run(vonNeumannMachine, getPT2Handlers(fifoQueue));
 
-    assertThat(fifoQueue.poll(), is(999));
+    assertThat(fifoQueue.poll(), is(999L));
   }
 
 
-  private List<MoveHandler> getPT2Handlers(LinkedList<Integer> fifoqueue) {
+  private List<MoveHandler> getPT2Handlers(LinkedList<Long> fifoqueue) {
     return asList(new SumHandler(),
         new MultiplicationHandler(),
         new InputHandler(fifoqueue),
